rf from_learn_menu_risk_management_guide
